requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
Pillow==10.2.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1